made simple test work, however...
this set of functions looks at a file systems as a disk-based one, not as one that could be dynamically
build and managed like those provided by QNX resource managers.  So, this is a fragile fix.  If a user
needs more, they'll have to fix/write themselves.  But the source is available!